Tweak SDL window creation params
These are still very sketchy. At some point I'll need to deal with
resolution properly. I'd like to turn on SDL_WINDOW_RESIZABLE but I'm
running into terrible problems with the implications of that and am
going to kick it down the road to when the engine is more mature
overall.