Run minicomp shader + present every frame
This is still pretty sketchy/hacky. In the long term I think I'd like
minicomp_state to be an instance of a proper class that can somehow set
up its Vulkan parts properly based on abstract input (perhaps the
mythical Game).

If Vulkan::minicomp_recreate_swch() actually gets called, horrible
things occur in my environment. From the log output it appears that
everything works properly, but the image in the window stops updating,
and the application stops responding to outside events (except very
slowly, sometimes). If I run it from within GDB and kill it from there,
it seems to be stuck hanging on the shader/copy Device::submit() fence,
at least most of the time. Sometimes I've been able to get it to close
while piping log output into less by sending the window close event to
SDL, then changing focus to the less terminal and scrolling to the
bottom to fetch the latest from the buffer. When this has worked, I'll
see that the last frame took several seconds (like 3). I imagine this is
all the result of some mistake I'm making, although I've been trying all
sorts of possible solutions for hours to no avail.

As long as I don't allow the window to be resized, though, the swapchain
never seems to require recreation, so I'm deciding to postpone
implementing dynamic changes in resolution for the time being. My hope
is that when I need to return to it, either:

    a. I'll be experienced enough with Vulkan to trivially spot my
       hypothetical mistake,
    b. I'll have improved the diagnostic features to make it easy to
       diagnose that way,
    c. it's actually due to a driver bug or something and will have been
       fixed long ago, or
    d. the engine will have changed around it in a way that makes the
       problem magically disappear or become irrelevant.

Fingers crossed.