Send elapsed time to minicomp shader
Facilitates animation; I have the inner blue part slowly pulsing now.
Obviously the shader itself is still in a great state of flux and
something of a mess, but I want to capture it in its current state
anyway, partially for my own future reference and partially just because
it works.