Refocus ShaderModule on binary data
In order to support shader data stored in the binary in addition to data
in an outside file. ShaderModule no longer stores a file as it's a poor
fit for this new scheme (and it doesn't appear to be necessary for the
outside data to persist for the life of the ShaderModule anyway).